✓

+++ média.java
@@ -7,16 +7,16 @@ public class média {
     public static void main(String[] args) {
         try (Scanner scan = new Scanner(System.in)){
 
-            double a, b, c, media;
+            float a, b, c, media;
             
                 System.out.print("Digite a primeira nota: ");
-                a = scan.nextInt();
+                a = scan.nextFloat();
             
                 System.out.println("Digite a segunda nota: ");
-                b = scan.nextInt();
+                b = scan.nextFloat();
             
                 System.out.println("Digite a terceira nota: ");
-                c = scan.nextInt();
+                c = scan.nextFloat();
 
                 media = (a+b+c)/3;
                 System.out.println("A média é: " + media);
